requirements.txt
flask
pyjwt
python-dotenv
bcrypt
pysqlite3
